Split the classes into their own files
