+++ lib/event-bus/src/rabbit-event-bus/Serializer.ts
@@ -1,0 +1,28 @@
+import { Event, EventIdentifier } from '../event-bus';
+import { MessageWrapper } from './types';
+
+export class Serialization {
+  // Maybe merge this class with AMQPConnector
+  public static defToExchange(def: EventIdentifier) {
+    return `event__${def.kind}-${def.namespace}`;
+  }
+
+  public static defToQueue(def: EventIdentifier, serviceName: string) {
+    return `consumer__${def.kind}-${def.namespace}__${serviceName}`;
+  }
+
+  public static eventToMessage<T extends object>(
+    event: Event<T>,
+  ): MessageWrapper<Event<T>> {
+    // Wrap the event in some internal transport layer format, in effect transforming the
+    // event into a message
+    return {
+      event,
+      meta: {
+        attempts: 0, // increments each process
+        retries: 10, // total retries
+        failures: 0, // increments each failure
+      },
+    };
+  }
+}
